move urlCheck to urlHandling
